requirements.txt
h5py
tensorboard
scikit-spatial
numpy==1.23.0
Pillow==9.5.0
ninja
pytorch-lightning~=1.1.7
torch_optimizer
opencv-python
seaborn
imageio
pyod
pycocotools
Cython
matplotlib
zivid
PyYAML
setuptools==59.5.0
torch==1.10.1+cu111
torchvision==0.11.2+cu111
torchaudio==0.10.1
open3d
scikit-learn